Support for tweeting library files independently of beets
 - Search library folder structure to compile dictionary of artists
 - Compare user input to artists and allow choice of closest match
 - List albums available
 - Get metadata and album artwork from relevant album folder

This update makes the script independent of the beets library,
however it still requires users to have a structured way of storing
their music (and is specifically setup for my personal beets structure).
The most useful application of this feature is tweeting album art from
albums stored on a NAS etc. which may not be linked to the beets
library on the user's PC.